requirements.txt
google-genai
torch
pydicom
torch>=1
torchvision>=0.5
scikit-image>=0.16
tqdm>=4
numpy>=1
pandas>=1
requests>=1
pillow>=5.3.0
imageio
python-dotenv
matplotlib